requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
django-filter==25.1
python-decouple==3.8
psycopg2-binary==2.9.10
gunicorn==22.0.0
redis==5.0.4
django-redis==5.4.0